requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
aiosmtplib
email-validator
python-dotenv
